auth_service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.1
email-validator==2.2.0
